tritlas/requirements.txt
amqp==1.4.9
anyjson==0.3.3
beautifulsoup4==4.9.1
billiard==3.3.0.23
bleach==3.2.1
celery==3.1.26.post2
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.17.0
dj-database-url==0.5.0
Django==2.2.9
django-celery==3.3.1
django-cloudinary-storage==0.2.3
django-filter==2.4.0
django-google-analytics-app==4.4.2
django-heroku==0.3.1
django-markdownify==0.8.1
djangorestframework==3.12.1
gunicorn==20.0.4
idna==2.10
kombu==3.0.37
Markdown==3.2.2
mock==4.0.2
packaging==20.4
Pillow==7.0.0
psycopg2==2.8.5
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
structlog==20.1.0
urllib3==1.25.10
webencodings==0.5.1
whitenoise==5.2.0
